requirements.txt
PyYAML==6.0.2
python-telegram-bot==22.6
python-dotenv==1.1.1
pillow==11.3.0